Annotations/testsetParser/requirements.txt
reflex>=0.3.6
python-dotenv>=1.0.0
huggingface-hub>=0.19.4
datasets>=2.15.0
pandas>=2.1.4
